Put TestEnvironment in its own file
