+++ dsl/test/interpreter/TestDSLInterpreter.java
@@ -17,7 +17,6 @@ import parser.AST.Node;
 import runtime.*;
 
 import semanticAnalysis.Scope;
-import semanticAnalysis.Symbol;
 import semanticAnalysis.SymbolTableParser;
 import semanticAnalysis.types.*;
 
@@ -209,31 +208,6 @@ public class TestDSLInterpreter {
 
     @DSLType(name = "quest_config")
     public record CustomQuestConfig(@DSLTypeMember Entity entity) {}
-
-    class TestEnvironment extends GameEnvironment {
-        public TestEnvironment() {
-            super();
-        }
-
-        @Override
-        protected void bindBuiltIns() {
-            for (IType type : BUILT_IN_TYPES) {
-                // load custom QuestConfig
-                if (!type.getName().equals("quest_config")
-                        && !type.getName().equals("game_object")) {
-                    globalScope.bind((Symbol) type);
-                }
-            }
-
-            var questConfigType =
-                    this.getTypeBuilder().createTypeFromClass(Scope.NULL, CustomQuestConfig.class);
-            loadTypes(new semanticAnalysis.types.IType[] {questConfigType});
-
-            for (Symbol func : NATIVE_FUNCTIONS) {
-                globalScope.bind(func);
-            }
-        }
-    }
 
     @Test
     public void aggregateTypeInstancing() {
